General cleanup and code refactor
