+++ build.gradle
@@ -1,10 +1,10 @@
 plugins {
     id 'java'
-    id 'org.jetbrains.intellij' version '0.4.14'
+    id 'org.jetbrains.intellij' version '0.4.16'
 }
 
 group 'uk.co.hadoopathome.intellij.avro'
-version '1.0.9'
+version '1.0.10'
 
 repositories {
     mavenCentral()
