+++ src/main/java/uk/co/hadoopathome/intellij/avro/fileformat/Reader.java
@@ -1,0 +1,10 @@
+package uk.co.hadoopathome.intellij.avro.fileformat;
+
+import java.util.List;
+
+public interface Reader {
+
+  List<String> getRecords(int numRecords);
+
+  String getSchema();
+}
