+++ src/test/java/uk/co/hadoopathome/intellij/avro/fileformat/AvroReaderTest.java
@@ -1,4 +1,4 @@
-package uk.co.hadoopathome.intellij.avro;
+package uk.co.hadoopathome.intellij.avro.fileformat;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -7,17 +7,18 @@ import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import org.junit.Test;
+import uk.co.hadoopathome.intellij.avro.fileformat.AvroReader;
+import uk.co.hadoopathome.intellij.avro.fileformat.Reader;
 
 public class AvroReaderTest {
   private static final String TWITTER_AVRO_FILE = "twitter.avro";
-  // From
   // https://github.com/apache/incubator-gobblin/blob/master/gobblin-core/src/test/resources/converter/pickfields_nested_with_union.avro
   private static final String COMPLEX_AVRO = "pickfields_nested_with_union.avro";
 
   @Test
   public void testGetSchema() throws IOException {
     File file = new File(getClass().getClassLoader().getResource(TWITTER_AVRO_FILE).getFile());
-    AvroReader avroReader = new AvroReader(file);
+    Reader avroReader = new AvroReader(file);
     String schema = avroReader.getSchema();
     assertTrue(schema.contains("A basic schema for storing Twitter messages"));
   }
@@ -25,7 +26,7 @@ public class AvroReaderTest {
   @Test
   public void testGetRecords() throws IOException {
     File file = new File(getClass().getClassLoader().getResource(TWITTER_AVRO_FILE).getFile());
-    AvroReader avroReader = new AvroReader(file);
+    Reader avroReader = new AvroReader(file);
     List<String> records = avroReader.getRecords(100);
     assertEquals(2, records.size());
     String firstRecord = records.get(0);
@@ -35,7 +36,7 @@ public class AvroReaderTest {
   @Test
   public void testComplexNesting() throws IOException {
     File file = new File(getClass().getClassLoader().getResource(COMPLEX_AVRO).getFile());
-    AvroReader avroReader = new AvroReader(file);
+    Reader avroReader = new AvroReader(file);
     List<String> records = avroReader.getRecords(100);
     assertEquals(6, records.size());
     String firstRecord = records.get(0);
