+++ src/test/java/uk/co/hadoopathome/intellij/avro/table/TableFormatterTest.java
@@ -1,4 +1,4 @@
-package uk.co.hadoopathome.intellij.avro;
+package uk.co.hadoopathome.intellij.avro.table;
 
 import static org.junit.Assert.assertEquals;
 
@@ -7,6 +7,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import org.junit.Test;
+import uk.co.hadoopathome.intellij.avro.table.TableFormatter;
 
 public class TableFormatterTest {
   @Test
